Refactor changes
Levigo to Carebase
