+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "com.levigo.levigoapp"
+        applicationId "com.levigo.Carebase"
         minSdkVersion 24
         targetSdkVersion 29
         versionCode 1
