+++ app/src/androidTest/java/com/levigo/Carebase/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.content.Context;
