+++ app/src/main/java/com/levigo/Carebase/CategoryAdapter.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.content.Context;
 import android.content.Intent;
