+++ app/src/main/java/com/levigo/Carebase/FilterActivity.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.content.Intent;
 import android.os.Bundle;
