+++ app/src/main/java/com/levigo/Carebase/FilterAdapter.java
@@ -1,6 +1,6 @@
 
 
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.content.Context;
 import android.util.Log;
