+++ app/src/main/java/com/levigo/Carebase/InventoryTemplate.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 public class InventoryTemplate {
