+++ app/src/main/java/com/levigo/Carebase/InventoryViewAdapter.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.util.Log;
 import android.view.LayoutInflater;
