+++ app/src/main/java/com/levigo/Carebase/MyFirebaseMessagingService.java
@@ -1,8 +1,5 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
-import android.app.Service;
-import android.content.Intent;
-import android.os.IBinder;
 import android.util.Log;
 
 import androidx.core.app.NotificationCompat;
