+++ app/src/main/java/com/levigo/Carebase/NetworkActivity.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import android.app.Activity;
 import android.content.Intent;
