+++ app/src/main/java/com/levigo/Carebase/UDIAdapter.java
@@ -1,6 +1,5 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
-import android.app.Activity;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,7 +9,6 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import java.util.List;
 import java.util.Map;
 
 public class UDIAdapter extends RecyclerView.Adapter<UDIAdapter.UDIHolder> {
