+++ app/src/test/java/com/levigo/Carebase/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.levigo.levigoapp;
+package com.levigo.Carebase;
 
 import org.junit.Test;
