+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='levigoapp'
+rootProject.name='Carebase'
 include ':app'
